fundamental:record: use int64_t for timestamp_delta
Kafka DefaultRecord.java and, librdkafka and other implementations
rely on the value having enough bits to store a long (int64_t)

public class DefaultRecord implements Record {

    private final int sizeInBytes;
    private final byte attributes;
    private final long offset;
    private final long timestamp;
    private final int sequence;
    private final ByteBuffer key;
    private final ByteBuffer value;
    private final Header[] headers;

}

We leave `long offset` as int32_t for now since this is unrelated
to the crash of test=0054 in librdkafka

see [ch934]

+++ src/v/model/record.h
@@ -100,6 +100,24 @@ private:
     iobuf _value;
 };
 
+/// \brief
+// DefaultRecord(int sizeInBytes,
+//               byte attributes,
+//               long offset,
+//               long timestamp,
+//               int sequence,
+//               ByteBuffer key,
+//               ByteBuffer value,
+//               Header[] headers) {
+//     this.sizeInBytes = sizeInBytes;
+//     this.attributes = attributes;
+//     this.offset = offset;
+//     this.timestamp = timestamp;
+//     this.sequence = sequence;
+//     this.key = key;
+//     this.value = value;
+//     this.headers = headers;
+// }
 class record {
 public:
     record() = default;
@@ -111,7 +129,7 @@ public:
     record(
       int32_t size_bytes,
       record_attributes attributes,
-      int32_t timestamp_delta,
+      int64_t timestamp_delta,
       int32_t offset_delta,
       int32_t key_size,
       iobuf key,
@@ -146,7 +164,7 @@ public:
 
     record_attributes attributes() const { return _attributes; }
 
-    int32_t timestamp_delta() const { return _timestamp_delta; }
+    int64_t timestamp_delta() const { return _timestamp_delta; }
 
     int32_t offset_delta() const { return _offset_delta; }
 
@@ -211,7 +229,7 @@ public:
 private:
     int32_t _size_bytes{0};
     record_attributes _attributes;
-    int32_t _timestamp_delta{0};
+    int64_t _timestamp_delta{0};
     int32_t _offset_delta{0};
     int32_t _key_size{0};
     iobuf _key;
